feat: new endpoint: routes

+++ internal/model/route.go
@@ -2,6 +2,6 @@ package model
 
 type Route struct {
 	BusID  int64 `db:"bus_id"`
-	StopID int   `db:"stop_id"`
+	StopID int64 `db:"stop_id"`
 	Step   uint8 `db:"step"`
 }
